Allow the setting of weights in metadata to influence sorting
By default discopower sorts the display alphabetically by English
name. This is good where there are lots of identity providers, but where
there are smaller numbers (eg social providers) it may be desirable to
override the sort order to influence user behaviour.

This introduces a 'discopower.weight' to all entries to be
weighted so that they appear higher or lower than they normally would.