Fix tab translations for new translation system.
First see if we have a translation for the full translation tag name.
If not (the translation attempt returns the full tag name untranslated),
fall back to see if we have something for only the tag name. This is
at the same time the final fallback - if we don't have that etither,
we display the bare tag name itself.

Fixes #3